update messaging information and item objects

+++ src/api/services/ListingItemService.ts
@@ -21,6 +21,8 @@ import { ItemInformationUpdateRequest } from '../requests/ItemInformationUpdateR
 import { PaymentInformationCreateRequest } from '../requests/PaymentInformationCreateRequest';
 import { PaymentInformationUpdateRequest } from '../requests/PaymentInformationUpdateRequest';
 import { MessagingInformationCreateRequest } from '../requests/MessagingInformationCreateRequest';
+import { MessagingInformationUpdateRequest } from '../requests/MessagingInformationUpdateRequest';
+
 import { ListingItemPostRequest } from '../requests/ListingItemPostRequest';
 import { ListingItemObjectCreateRequest } from '../requests/ListingItemObjectCreateRequest';
 import { ListingItemObjectUpdateRequest } from '../requests/ListingItemObjectUpdateRequest';
@@ -203,31 +205,58 @@ export class ListingItemService {
         }
 
         // find related record and delete it and recreate related data
-        let messagingInformation = updatedListingItem.related('MessagingInformation').toJSON() || [];
-        for (const msgInfo of messagingInformation) {
-            await this.messagingInformationService.destroy(msgInfo.id);
+        const existintMessagingInformation = updatedListingItem.related('MessagingInformation').toJSON() || [];
+
+        const newMessagingInformation = body.messagingInformation || [];
+
+        // delete MessagingInformation if not exist with new params
+        for (const msgInfo of existintMessagingInformation) {
+            if (!await this.checkExistingObject(newMessagingInformation, 'publicKey', msgInfo.publicKey)) {
+                await this.messagingInformationService.destroy(msgInfo.id);
+            }
         }
-        messagingInformation = body.messagingInformation || [];
-        for (const msgInfo of messagingInformation) {
+
+        // update or create messaging itemInformation
+        for (const msgInfo of newMessagingInformation) {
             msgInfo.listing_item_id = id;
-            await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+            const message = await this.checkExistingObject(existintMessagingInformation, 'publicKey', msgInfo.publicKey);
+            if (message) {
+                message.protocol = msgInfo.protocol;
+                message.publicKey = msgInfo.publicKey;
+                await this.messagingInformationService.update(message.id, msgInfo as MessagingInformationUpdateRequest);
+            } else {
+                await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+            }
         }
 
-        // find related record and delete it and recreate related data
-        let listingItemObjects = updatedListingItem.related('ListingItemObjects').toJSON() || [];
+        const newListingItemObjects = body.listingItemObjects || [];
 
-        for (const object of listingItemObjects) {
+        // find related listingItemObjects
+        const existingListingItemObjects = updatedListingItem.related('ListingItemObjects').toJSON() || [];
+
+        // find highestOrderNumber
+        const highestOrderNumber = await this.findHighesOrderNumber(newListingItemObjects);
+
+        const objectsToBeUpdated = [] as any;
+        for (const object of existingListingItemObjects) {
+            // check if order number is greter than highestOrderNumber then delete
+            if (object.order > highestOrderNumber) {
+                await this.listingItemObjectService.destroy(object.id);
+            } else {
+                objectsToBeUpdated.push(object);
+            }
+        }
+        // create or update listingItemObjects
+        for (const object of newListingItemObjects) {
             object.listing_item_id = id;
-            await this.listingItemObjectService.destroy(object.id);
-        }
+            const itemObject = await this.checkExistingObject(objectsToBeUpdated, 'order', object.order);
 
-        // add new
-        listingItemObjects = body.listingItemObjects || [];
-        for (const object of listingItemObjects) {
-            object.listing_item_id = id;
-            await this.listingItemObjectService.create(object as ListingItemObjectCreateRequest);
+            if (itemObject) {
+                await this.listingItemObjectService.update(itemObject.id, object as ListingItemObjectUpdateRequest);
+            } else {
+                await this.listingItemObjectService.create(object as ListingItemObjectCreateRequest);
+            }
         }
-
 
         // finally find and return the updated listingItem
         return await this.findOne(id);
@@ -281,5 +310,20 @@ export class ListingItemService {
         this.messageBroadcastService.broadcast(addItemMessage as ListingItemMessage);
         return itemTemplate;
 
+    }
+
+    // check if object is exist in a array
+    private async checkExistingObject(objectArray: string[], fieldName: string, value: string | number): Promise<any> {
+        return await _.find(objectArray, (object) => {
+            return ( object[fieldName] === value );
+        });
+    }
+
+    // find highest order number from listingItemObjects
+    private async findHighesOrderNumber(listingItemObjects: string[]): Promise<any> {
+        const highestOrder = await _.maxBy(listingItemObjects, (itemObject) => {
+          return itemObject['order'];
+        });
+        return highestOrder ? highestOrder['order'] : 0;
     }
 }
