+++ src/api/services/ListingItemTemplateService.ts
@@ -156,32 +156,72 @@ export class ListingItemTemplateService {
             await this.paymentInformationService.destroy(paymentInformation.id);
         }
 
-        // find related record and delete it and recreate related data
-        let messagingInformation = updatedListingItemTemplate.related('MessagingInformation').toJSON() || [];
-        for (const msgInfo of messagingInformation) {
-            msgInfo.listing_item_template_id = id;
-            await this.messagingInformationService.destroy(msgInfo.id);
-        }
-        // add new
-        messagingInformation = body.messagingInformation || [];
-        for (const msgInfo of messagingInformation) {
-            msgInfo.listing_item_template_id = id;
-            await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
-        }
+        // // find related record and delete it and recreate related data
+        // let messagingInformation = updatedListingItemTemplate.related('MessagingInformation').toJSON() || [];
+        // for (const msgInfo of messagingInformation) {
+        //     msgInfo.listing_item_template_id = id;
+        //     await this.messagingInformationService.destroy(msgInfo.id);
+        // }
+        // // add new
+        // messagingInformation = body.messagingInformation || [];
+        // for (const msgInfo of messagingInformation) {
+        //     msgInfo.listing_item_template_id = id;
+        //     await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+        // }
 
         // find related record and delete it and recreate related data
-        let listingItemObjects = updatedListingItemTemplate.related('ListingItemObjects').toJSON() || [];
+        const existintMessagingInformation = updatedListingItemTemplate.related('MessagingInformation').toJSON() || [];
 
-        for (const object of listingItemObjects) {
-            object.listing_item_template_id = id;
-            await this.listingItemObjectService.destroy(object.id);
+        const newMessagingInformation = body.messagingInformation || [];
+
+        // delete MessagingInformation if not exist with new params
+        for (const msgInfo of existintMessagingInformation) {
+            if (!await this.checkExistingObject(newMessagingInformation, 'id', msgInfo.id)) {
+                await this.messagingInformationService.destroy(msgInfo.id);
+            }
         }
 
-        // add new
-        listingItemObjects = body.listingItemObjects || [];
-        for (const object of listingItemObjects) {
+        // update or create messaging itemInformation
+        for (const msgInfo of newMessagingInformation) {
+            msgInfo.listing_item_template_id = id;
+            const message = await this.checkExistingObject(existintMessagingInformation, 'id', msgInfo.id);
+            delete msgInfo.id;
+            if (message) {
+                message.protocol = msgInfo.protocol;
+                message.publicKey = msgInfo.publicKey;
+                await this.messagingInformationService.update(message.id, msgInfo as MessagingInformationUpdateRequest);
+            } else {
+                await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+            }
+        }
+
+        const newListingItemObjects = body.listingItemObjects || [];
+        // find related listingItemObjects
+        const existingListingItemObjects = updatedListingItemTemplate.related('ListingItemObjects').toJSON() || [];
+
+        // find highestOrderNumber
+        const highestOrderNumber = await this.findHighesOrderNumber(newListingItemObjects);
+
+        const objectsToBeUpdated = [] as any;
+        for (const object of existingListingItemObjects) {
+            // check if order number is greter than highestOrderNumber then delete
+            if (object.order > highestOrderNumber) {
+                await this.listingItemObjectService.destroy(object.id);
+            } else {
+                objectsToBeUpdated.push(object);
+            }
+        }
+
+        // create or update listingItemObjects
+        for (const object of newListingItemObjects) {
             object.listing_item_template_id = id;
-            await this.listingItemObjectService.create(object as ListingItemObjectCreateRequest);
+            const itemObject = await this.checkExistingObject(objectsToBeUpdated, 'order', object.order);
+
+            if (itemObject) {
+                await this.listingItemObjectService.update(itemObject.id, object as ListingItemObjectUpdateRequest);
+            } else {
+                await this.listingItemObjectService.create(object as ListingItemObjectCreateRequest);
+            }
         }
 
         // finally find and return the updated listingItem
@@ -203,4 +243,18 @@ export class ListingItemTemplateService {
 
     }
 
+    // check if object is exist in a array
+    private async checkExistingObject(objectArray: string[], fieldName: string, value: string | number): Promise<any> {
+        return await _.find(objectArray, (object) => {
+            return ( object[fieldName] === value );
+        });
+    }
+
+    // find highest order number from listingItemObjects
+    private async findHighesOrderNumber(listingItemObjects: string[]): Promise<any> {
+        const highestOrder = await _.maxBy(listingItemObjects, (itemObject) => {
+          return itemObject['order'];
+        });
+        return highestOrder ? highestOrder['order'] : 0;
+    }
 }
